+++ repository/src/db_diesel/mod.rs
@@ -7,7 +7,7 @@ mod filter_sort_pagination;
 mod invoice;
 mod invoice_line;
 mod invoice_line_row;
-mod invoice_query;
+mod invoice_row;
 mod item;
 mod item_query;
 mod key_value_store;
@@ -52,7 +52,7 @@ pub use filter_sort_pagination::*;
 pub use invoice::*;
 pub use invoice_line::*;
 pub use invoice_line_row::*;
-pub use invoice_query::*;
+pub use invoice_row::*;
 pub use item::ItemRepository;
 pub use item_query::*;
 pub use key_value_store::*;
