+++ repository/src/mock/mod.rs
@@ -79,7 +79,7 @@ use self::unit::mock_units;
 
 use super::{
     db_diesel::{
-        InvoiceRepository, ItemRepository, NameRepository, NameStoreJoinRepository,
+        InvoiceRowRepository, ItemRepository, NameRepository, NameStoreJoinRepository,
         StorageConnection, StoreRowRepository, UnitRowRepository,
     },
     schema::*,
@@ -438,7 +438,7 @@ pub async fn insert_mock_data(
         }
 
         if inserts.invoices {
-            let repo = InvoiceRepository::new(connection);
+            let repo = InvoiceRowRepository::new(connection);
             for row in &mock_data.invoices {
                 repo.upsert_one(&row).unwrap();
             }
