+++ service/src/invoice/inbound_shipment/insert/validate.rs
@@ -1,6 +1,6 @@
 use crate::validate::{check_other_party, CheckOtherPartyType, OtherPartyErrors};
 use repository::Name;
-use repository::{InvoiceRepository, RepositoryError, StorageConnection};
+use repository::{InvoiceRowRepository, RepositoryError, StorageConnection};
 
 use super::{InsertInboundShipment, InsertInboundShipmentError};
 
@@ -32,7 +32,7 @@ fn check_invoice_does_not_exists(
     id: &str,
     connection: &StorageConnection,
 ) -> Result<(), InsertInboundShipmentError> {
-    let result = InvoiceRepository::new(connection).find_one_by_id(id);
+    let result = InvoiceRowRepository::new(connection).find_one_by_id(id);
 
     if let Err(RepositoryError::NotFound) = &result {
         Ok(())
