+++ service/src/invoice/outbound_shipment/batch.rs
@@ -281,7 +281,7 @@ mod test {
             MockDataInserts,
         },
         test_db::setup_all,
-        InvoiceLineRowRepository, InvoiceRepository,
+        InvoiceLineRowRepository, InvoiceRowRepository,
     };
     use util::inline_init;
 
@@ -347,7 +347,7 @@ mod test {
         );
 
         assert_eq!(
-            InvoiceRepository::new(&connection)
+            InvoiceRowRepository::new(&connection)
                 .find_one_by_id_option("new_id")
                 .unwrap(),
             None
@@ -368,7 +368,7 @@ mod test {
             .unwrap();
 
         assert_ne!(
-            InvoiceRepository::new(&connection)
+            InvoiceRowRepository::new(&connection)
                 .find_one_by_id_option("new_id")
                 .unwrap(),
             None
